requirements.txt

telethon
python-telegram-bot
python-dotenv
